LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/computed/get-position-expected.txt


FAIL Computed position sticky even if it does not apply assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Computed position absolute even if it does not apply assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
